# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Darbby Platform API — Backend for customer and merchant apps
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listMyTrips
      tags: [trips]
      summary: List my trips
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED, CANCELLED]
      responses:
        "200":
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTripRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /offers/{id}/counter:
    post:
      operationId: counterOffer
      tags: [offers]
      summary: Submit a counter-offer (negotiation round)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferRequest"
      responses:
        "200":
          description: Counter-offer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negotiation"

  # ─── MERCHANT OFFERS ──────────────────────────────────────────────────────
  /merchant/trips:
    get:
      operationId: listActiveTripsMerchant
      tags: [merchants]
      summary: List active trips near merchant branches
      security:
        - bearerAuth: []
      parameters:
        - name: branchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Matching trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
  /merchant/trips/{tripId}/offer:
    post:
      operationId: sendOffer
      tags: [merchants]
      summary: Send an offer for a trip
      security:
        - bearerAuth: []
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOfferRequest"
      responses:
        "201":
          description: Offer sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
  /merchant/offers:
    get:
      operationId: listMerchantOffers
      tags: [merchants]
      summary: List merchant's sent offers
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Offers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedOffers"
  /merchant/offers/{id}/accept-counter:
    post:
      operationId: acceptCounter
      tags: [merchants]
      summary: Merchant accepts a customer counter-offer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          description: Counter accepted, offer finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDetail"
  /merchant/offers/{id}/counter:
    post:
      operationId: merchantCounter
      tags: [merchants]
      summary: Merchant sends a counter-offer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CounterOfferRequest"
      responses:
        "200":
          description: Counter submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Negotiation"

  # ─── MERCHANT PROFILE ─────────────────────────────────────────────────────
  /merchant/profile:
    get:
      operationId: getMerchantProfile
      tags: [merchants]
      summary: Get merchant profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Merchant profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProfile"
    put:
      operationId: updateMerchantProfile
      tags: [merchants]
      summary: Update merchant profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMerchantProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantProfile"

  # ─── BRANCHES ─────────────────────────────────────────────────────────────
  /merchant/branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List merchant branches
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantBranch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a new branch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantBranch"
  /merchant/branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantBranch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Deleted

  # ─── PRODUCTS ─────────────────────────────────────────────────────────────
  /merchant/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List merchant products
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /merchant/products/{id}:
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Deleted

  # ─── AUTO-NEGOTIATOR ──────────────────────────────────────────────────────
  /merchant/auto-negotiator:
    get:
      operationId: getAutoNegSettings
      tags: [auto-negotiator]
      summary: Get auto-negotiator settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoNegSettings"
    put:
      operationId: updateAutoNegSettings
      tags: [auto-negotiator]
      summary: Update auto-negotiator settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAutoNegRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoNegSettings"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────
...
  /merchant/stats:
    get:
      operationId: getMerchantStats
      tags: [merchants]
      summary: Get merchant dashboard stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MerchantStats"
  /merchant/commission:
    get:
      operationId: getMerchantCommission
      tags: [merchants]
      summary: Get commission ledger summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Commission summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionSummary"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    CreateTripRequest:
      type: object
      properties:
        title:
          type: string
        tripPurpose:
          type: string
          enum: [WORK, TOURISM, UMRAH, FAMILY_VISIT, MEDICAL, EDUCATION, OTHER]
        originName:
          type: string
        originLat:
          type: number
        originLng:
          type: number
        destinationName:
          type: string
        destLat:
          type: number
        destLng:
          type: number
        routePolyline:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        vehicleProfileId:
          type: string
        isPublic:
          type: boolean
      required: [originName, originLat, originLng, destinationName, destLat, destLng, departureTime, tripPurpose]
    Trip:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        vehicleProfileId:
          type: string
        title:
          type: string
        tripPurpose:
          type: string
        originName:
          type: string
        destinationName:
          type: string
        routePolyline:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        status:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, tripPurpose, originName, destinationName, departureTime, status]
...
    SendOfferRequest:
      type: object
      properties:
        branchId:
          type: string
        message:
          type: string
        expiresAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
            required: [productId, quantity, unitPrice]
      required: [items, expiresAt]
    CounterOfferRequest:
      type: object
      properties:
        proposedPrice:
          type: number
        message:
          type: string
      required: [proposedPrice]
    OfferDetail:
      type: object
      properties:
        id:
          type: string
        tripId:
          type: string
        merchantId:
          type: string
        branchId:
          type: string
        message:
          type: string
        totalPrice:
          type: number
        finalPrice:
          type: number
        status:
          type: string
        isAutoOffer:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OfferItemDetail"
        negotiations:
          type: array
          items:
            $ref: "#/components/schemas/Negotiation"
        createdAt:
          type: string
          format: date-time
      required: [id, tripId, merchantId, totalPrice, status]
...
    Negotiation:
      type: object
      properties:
        id:
          type: string
        offerId:
          type: string
        senderType:
          type: string
          enum: [USER, MERCHANT, SYSTEM]
        proposedPrice:
          type: number
        message:
          type: string
        isAuto:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, offerId, senderType, proposedPrice]
    PaginatedOffers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OfferDetail"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [data, total, page, pageSize]

    # ─── Merchant ────────────────────────────────────────────────────────────
    MerchantProfile:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        ownerName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        subscriptionPlan:
          type: string
        nationalAddress:
          type: string
        nafathVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, businessName, ownerName, email, phone, status]
    UpdateMerchantProfileRequest:
      type: object
      properties:
        businessName:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        nationalAddress:
          type: string
        fcmToken:
          type: string

    # ─── Branch ──────────────────────────────────────────────────────────────
    MerchantBranch:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        branchName:
          type: string
        branchCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        addressText:
          type: string
        serviceRadiusKm:
          type: number
        status:
          type: string
        isPrimary:
          type: boolean
        workingHours:
          type: object
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, merchantId, branchName, status, isPrimary]
    CreateBranchRequest:
      type: object
      properties:
        branchName:
          type: string
        branchCode:
          type: string
        lat:
          type: number
        lng:
          type: number
        addressText:
          type: string
        serviceRadiusKm:
          type: number
        isPrimary:
          type: boolean
        workingHours:
          type: object
        phone:
          type: string
      required: [branchName, lat, lng]

    # ─── Product ─────────────────────────────────────────────────────────────
...
    AutoNegSettings:
      type: object
      properties:
        id:
          type: string
        merchantId:
          type: string
        isEnabled:
          type: boolean
        responseDelayMin:
          type: integer
        purposeRules:
          type: object
        products:
          type: array
          items:
            $ref: "#/components/schemas/AutoNegProduct"
      required: [id, merchantId, isEnabled]
...